split out racing TestMonitorPass
